fix: change to user home directory before Oh My Zsh installation
The Oh My Zsh installer uses "cd -" to return to the previous directory after installation. When the bastion script is run from /home/debian with sudo, and then switches to the bastion user, the installer tries to cd back to /home/debian, which the bastion user cannot access.

Error was:
  /tmp/omz-install.sh: 327: cd: can't cd to /home/debian

This happened because:
1. Script runs from /home/debian (user's home directory)
2. sudo -u bastion env HOME=/home/bastion ... runs installer
3. Installer successfully cd to /home/bastion/.oh-my-zsh
4. Installer runs "cd -" to return to original directory
5. Original $PWD was /home/debian, which bastion user can't access

Solution:
  (cd "$USER_HOME" && sudo -u "$USERNAME" env ...)

The subshell (parentheses) changes to the target user's home directory before running sudo, so when the installer uses "cd -", it returns to the user's own home directory, which they can access.

This is the final fix needed for Oh My Zsh installation to work correctly when the bastion script is run by a different user than the target user.